move files to sub-directory
