+++ src/components/atoms/icon/Icon.module.css
